Moves rdpdr tests to their own file
